Split constraints into multiple files
