+++ src/models/NotNullConstraint.js
@@ -1,0 +1,9 @@
+export default class NotNullConstraint {
+  static get label() {
+    return 'NOT NULL';
+  }
+
+  toDDL() {
+    return this.constructor.label;
+  }
+}
