+++ src/models/NullConstraint.js
@@ -1,0 +1,9 @@
+export default class NullConstraint {
+  static get label() {
+    return 'NULL';
+  }
+
+  toDDL() {
+    return this.constructor.label;
+  }
+}
