+++ src/models/PrimaryKeyConstraint.js
@@ -1,0 +1,9 @@
+export default class PrimaryKeyConstraint {
+  static get label() {
+    return 'PRIMARY KEY';
+  }
+
+  toDDL() {
+    return this.constructor.label;
+  }
+}
